added player controls
added music player controls. Used hook useRef in order to select the html audio tag in the component to play/pause the audio file. Initial state for isPlaying state is false. When play button is clicked the playSong function runs and toggles the state value. When false the audio.current value is pause and play when true. The font awesome icon also toggles when clicked.