added skip and play previous logic
Takes the current song id and checks to see which key in the array has a matching id. This determines the index of the song currently selected. If they click on skip-forward increase the index by 1 to go to the next song. If song is skipped back is moves back in the array. Also, if the audio is playing it will continue to play.