added library
the library becomes active when the librarystatus state changes to true by clicking on the button in the navbar. When the song is clicked in the library it will become the new selected song. If the song was playing the new song will continue to play. The song that has true as the value for the active key will have a colored background to show it's selected.